[Gold V] Title: 용액, Time: 264 ms, Memory: 32440 KB -BaekjoonHub

+++ 백준/Gold/2467. 용액/용액.java
@@ -19,7 +19,7 @@ public class Main{
         int start = 0;
         int end = N - 1;
         while (start < end){
-            long sum = arr[start] + arr[end];
+            int sum = arr[start] + arr[end];
             if(sum == 0) {
                 ldx = start;
                 rdx = end;
